fixed few problems with typescript and converted more js files to tsx
